+++ src/main/java/com/woowacourse/caffeine/application/MenuItemService.java
@@ -1,5 +1,6 @@
 package com.woowacourse.caffeine.application;
 
+import com.woowacourse.caffeine.application.object.MenuItemResponse;
 import com.woowacourse.caffeine.domain.MenuItem;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
