+++ src/main/java/com/woowacourse/caffeine/application/ShopService.java
@@ -1,5 +1,7 @@
 package com.woowacourse.caffeine.application;
 
+import com.woowacourse.caffeine.application.object.ShopCreateRequest;
+import com.woowacourse.caffeine.application.object.ShopResponse;
 import com.woowacourse.caffeine.domain.Shop;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
