+++ src/main/java/com/woowacourse/caffeine/application/object/ShopResponse.java
@@ -1,4 +1,4 @@
-package com.woowacourse.caffeine.application;
+package com.woowacourse.caffeine.application.object;
 
 public class ShopResponse {
     public final long id;
